Add UserProvider
